// src/Api/Types/RecipeItemDetails.ts
export interface RecipeItemDetails{
    recipeId: string;
    recipeMeal: string;
    recipeDrinkAlternative?: string;
    recipeCategory: string;
    recipeArea: string;
    recipeThumb: string;
    recipeTags?: string;
    recipeYoutube: string;
    recipeInstructions: string;
    recipeIngredient1: string;
    recipeIngredient2: string;
    recipeIngredient3: string;
    recipeIngredient4: string;
    recipeIngredient5: string;
    recipeIngredient6: string;
    recipeIngredient7: string;
    recipeIngredient8: string;
    recipeIngredient9: string;
    recipeIngredient10: string;
    recipeIngredient11: string;
    recipeIngredient12: string;
    recipeIngredient13: string;
    recipeIngredient14: string;
    recipeIngredient15: string;
    recipeIngredient16: string;
    recipeIngredient17: string;
    recipeIngredient18: string;
    recipeIngredient19: string;
    recipeIngredient20: string;
    recipeMeasure1: string;
    recipeMeasure2: string;
    recipeMeasure3: string;
    recipeMeasure4: string;
    recipeMeasure5: string;
    recipeMeasure6: string;
    recipeMeasure7: string;
    recipeMeasure8: string;
    recipeMeasure9: string;
    recipeMeasure10: string;
    recipeMeasure11: string;
    recipeMeasure12: string;
    recipeMeasure13: string;
    recipeMeasure14: string;
    recipeMeasure15: string;
    recipeMeasure16: string;
    recipeMeasure17: string;
    recipeMeasure18: string;
    recipeMeasure19: string;
    recipeMeasure20: string;
}